+++ synapse-kafka/src/test/java/de/otto/synapse/endpoint/receiver/kafka/KafkaMessageLogReceiverEndpointIntegrationTest.java
@@ -2,20 +2,20 @@ package de.otto.synapse.endpoint.receiver.kafka;
 
 import de.otto.synapse.channel.ChannelPosition;
 import de.otto.synapse.channel.ShardPosition;
-import de.otto.synapse.channel.ShardResponse;
 import de.otto.synapse.channel.StartFrom;
 import de.otto.synapse.consumer.MessageConsumer;
 import de.otto.synapse.endpoint.MessageInterceptorRegistry;
+import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.apache.kafka.clients.producer.ProducerConfig;
 import org.apache.kafka.clients.producer.ProducerRecord;
+import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.serialization.StringDeserializer;
 import org.apache.kafka.common.serialization.StringSerializer;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.context.annotation.Bean;
@@ -33,41 +33,39 @@ import java.util.*;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
-import java.util.function.Predicate;
+import java.util.stream.Collectors;
 
 import static com.google.common.collect.Sets.newConcurrentHashSet;
 import static de.otto.synapse.channel.ChannelPosition.channelPosition;
 import static de.otto.synapse.channel.ChannelPosition.fromHorizon;
 import static de.otto.synapse.channel.ShardPosition.fromPosition;
-import static de.otto.synapse.channel.StopCondition.emptyResponse;
+import static de.otto.synapse.channel.StopCondition.endOfChannel;
 import static java.util.concurrent.Executors.newFixedThreadPool;
 import static org.awaitility.Awaitility.waitAtMost;
 import static org.awaitility.Duration.TEN_SECONDS;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.*;
-import static org.slf4j.LoggerFactory.getLogger;
 import static org.springframework.kafka.test.utils.KafkaTestUtils.consumerProps;
 import static org.springframework.kafka.test.utils.KafkaTestUtils.producerProps;
 
 @RunWith(SpringRunner.class)
 @DirtiesContext
 @EmbeddedKafka(
-        partitions = 1,
+        partitions = 2,
         topics = KafkaMessageLogReceiverEndpointIntegrationTest.KAFKA_TOPIC)
-    public class KafkaMessageLogReceiverEndpointIntegrationTest {
-    private static final Logger LOG = getLogger(KafkaMessageLogReceiverEndpointIntegrationTest.class);
+public class KafkaMessageLogReceiverEndpointIntegrationTest {
     public static final String KAFKA_TOPIC = "test-stream";
 
     @Autowired
     private KafkaTemplate<String, String> template;
+
     @Autowired
-    ApplicationEventPublisher eventPublisher;
+    private ApplicationEventPublisher eventPublisher;
 
     @Autowired
     private EmbeddedKafkaBroker embeddedKafkaBroker;
 
     private KafkaConsumer<String, String> kafkaConsumer;
-    private MessageInterceptorRegistry interceptorRegistry;
     private KafkaMessageLogReceiverEndpoint endpoint;
     private ExecutorService executorService = newFixedThreadPool(2);
 
@@ -78,9 +76,10 @@ import static org.springframework.kafka.test.utils.KafkaTestUtils.producerProps;
                 groupId,
                 "true",
                 embeddedKafkaBroker));
+        configs.put(ConsumerConfig.MAX_PARTITION_FETCH_BYTES_CONFIG, 10); // set to small value to enforce multiple reads from kafka
         kafkaConsumer = new KafkaConsumer<>(configs, new StringDeserializer(), new StringDeserializer());
 
-        interceptorRegistry = new MessageInterceptorRegistry();
+        MessageInterceptorRegistry interceptorRegistry = new MessageInterceptorRegistry();
         endpoint = new KafkaMessageLogReceiverEndpoint(
                 KAFKA_TOPIC,
                 interceptorRegistry,
@@ -101,7 +100,7 @@ import static org.springframework.kafka.test.utils.KafkaTestUtils.producerProps;
         template.send(new ProducerRecord<>(KAFKA_TOPIC, "my-aggregate-id", "{\"event\":\"Test Event\"}"));
 
         // when
-        final CompletableFuture<ChannelPosition> channelPosition = endpoint.consumeUntil(fromHorizon(), messagesReceived());
+        final CompletableFuture<ChannelPosition> channelPosition = endpoint.consumeUntil(fromHorizon(), endOfChannel());
 
         // then
         waitAtMost(TEN_SECONDS).until(channelPosition::isDone);
@@ -114,7 +113,7 @@ import static org.springframework.kafka.test.utils.KafkaTestUtils.producerProps;
         template.send(new ProducerRecord<>(KAFKA_TOPIC, "my-aggregate-id", "{\"event\":\"Test Event\"}"));
 
         // when
-        final CompletableFuture<ChannelPosition> channelPosition = endpoint.consumeUntil(fromHorizon(), messagesReceived());
+        final CompletableFuture<ChannelPosition> channelPosition = endpoint.consumeUntil(fromHorizon(), endOfChannel());
 
         // then
         waitAtMost(TEN_SECONDS).until(channelPosition::isDone);
@@ -129,7 +128,7 @@ import static org.springframework.kafka.test.utils.KafkaTestUtils.producerProps;
         // when
         final Set<String> receivedMessageKeys = newConcurrentHashSet();
         endpoint.getMessageDispatcher().add(MessageConsumer.of(".*", String.class, m -> receivedMessageKeys.add(m.getKey().compactionKey())));
-        final CompletableFuture<ChannelPosition> channelPosition = endpoint.consumeUntil(fromHorizon(), messagesReceived());
+        final CompletableFuture<ChannelPosition> channelPosition = endpoint.consumeUntil(fromHorizon(), endOfChannel());
 
         // then
         waitAtMost(TEN_SECONDS).until(channelPosition::isDone);
@@ -146,16 +145,16 @@ import static org.springframework.kafka.test.utils.KafkaTestUtils.producerProps;
             template.send(new ProducerRecord<>(KAFKA_TOPIC, "shouldStartFromHorizon" + i, "{\"event\":\"Test Event\"}"));
         }
 
-        endpoint.consumeUntil(fromHorizon(), emptyResponse()).join();
+        endpoint.consumeUntil(fromHorizon(), endOfChannel()).join();
 
         // when
         final List<String> receivedMessageKeys = new ArrayList<>();
         endpoint.getMessageDispatcher().add(MessageConsumer.of("shouldStartFromHorizon.*", String.class, m -> receivedMessageKeys.add(m.getKey().compactionKey())));
-        final CompletableFuture<ChannelPosition> channelPosition = endpoint.consumeUntil(fromHorizon(), messagesReceived());
+        final CompletableFuture<ChannelPosition> channelPosition = endpoint.consumeUntil(fromHorizon(), endOfChannel());
 
         // then
         waitAtMost(TEN_SECONDS).until(channelPosition::isDone);
-        assertThat(receivedMessageKeys.subList(receivedMessageKeys.size()-10, receivedMessageKeys.size()), contains("shouldStartFromHorizon0", "shouldStartFromHorizon1", "shouldStartFromHorizon2", "shouldStartFromHorizon3", "shouldStartFromHorizon4", "shouldStartFromHorizon5", "shouldStartFromHorizon6", "shouldStartFromHorizon7", "shouldStartFromHorizon8", "shouldStartFromHorizon9"));
+        assertThat(receivedMessageKeys.subList(receivedMessageKeys.size() - 10, receivedMessageKeys.size()), containsInAnyOrder("shouldStartFromHorizon0", "shouldStartFromHorizon1", "shouldStartFromHorizon2", "shouldStartFromHorizon3", "shouldStartFromHorizon4", "shouldStartFromHorizon5", "shouldStartFromHorizon6", "shouldStartFromHorizon7", "shouldStartFromHorizon8", "shouldStartFromHorizon9"));
     }
 
     @Test
@@ -165,18 +164,18 @@ import static org.springframework.kafka.test.utils.KafkaTestUtils.producerProps;
             template.send(new ProducerRecord<>(KAFKA_TOPIC, "" + i, "{\"event\":\"Test Event\"}"));
         }
 
-        endpoint.consumeUntil(fromHorizon(), emptyResponse()).join();
+        endpoint.consumeUntil(fromHorizon(), endOfChannel()).join();
 
         // when
         final Set<String> receivedMessageKeys = newConcurrentHashSet();
         endpoint.getMessageDispatcher().add(MessageConsumer.of(".*", String.class, m -> receivedMessageKeys.add(m.getKey().compactionKey())));
 
         final ChannelPosition startFrom = channelPosition(fromPosition("0", "25"), fromPosition("1", "25"));
-        final CompletableFuture<ChannelPosition> channelPosition = endpoint.consumeUntil(startFrom, messagesReceived());
+        final CompletableFuture<ChannelPosition> channelPosition = endpoint.consumeUntil(startFrom, endOfChannel());
 
         // then
         waitAtMost(TEN_SECONDS).until(channelPosition::isDone);
-        assertThat(receivedMessageKeys.size(), is(greaterThan(40)));
+        assertThat(receivedMessageKeys.size(), is(48));
     }
 
     @Test
@@ -186,7 +185,7 @@ import static org.springframework.kafka.test.utils.KafkaTestUtils.producerProps;
             template.send(new ProducerRecord<>(KAFKA_TOPIC, "" + i, "{\"event\":\"Test Event\"}")).get();
         }
 
-        ChannelPosition startFrom = endpoint.consumeUntil(fromHorizon(), emptyResponse()).join();
+        ChannelPosition startFrom = endpoint.consumeUntil(fromHorizon(), endOfChannel()).join();
 
         // when
         for (int j = 0; j < 10; j++) {
@@ -202,13 +201,28 @@ import static org.springframework.kafka.test.utils.KafkaTestUtils.producerProps;
             // then
             waitAtMost(TEN_SECONDS).until(channelPosition::isDone);
             assertThat(receivedMessageKeys.size(), is(10));
-            assertThat(receivedMessageKeys, contains("continued-"+j + "/0", "continued-"+j + "/1", "continued-"+j + "/2", "continued-"+j + "/3", "continued-"+j + "/4", "continued-"+j + "/5", "continued-"+j + "/6", "continued-"+j + "/7", "continued-"+j + "/8", "continued-"+j + "/9"));
+            assertThat(receivedMessageKeys, containsInAnyOrder("continued-" + j + "/0", "continued-" + j + "/1", "continued-" + j + "/2", "continued-" + j + "/3", "continued-" + j + "/4", "continued-" + j + "/5", "continued-" + j + "/6", "continued-" + j + "/7", "continued-" + j + "/8", "continued-" + j + "/9"));
             startFrom = channelPosition.get();
         }
     }
 
-    private Predicate<ShardResponse> messagesReceived() {
-        return shardResponse -> !shardResponse.getMessages().isEmpty();
+    @Test
+    public void shouldConsumeUntilBothPartitionsAreAtEnd() throws ExecutionException, InterruptedException {
+        // given
+        for (int i = 0; i < 5; i++) {  //send to partition 1 and 2 alternating
+            template.send(new ProducerRecord<>(KAFKA_TOPIC, i % 2, "" + i, "{\"event\":\"Test Event\"}")).get();
+        }
+        for (int i = 10; i < 20; i++) { //only send to partition 1, so that this partition has more entries than the other
+            template.send(new ProducerRecord<>(KAFKA_TOPIC, 1, "" + i, "{\"event\":\"Test Event\"}")).get();
+        }
+
+        //when
+        ChannelPosition channelPosition = endpoint.consumeUntil(fromHorizon(), endOfChannel()).join();
+
+        //then
+        List<TopicPartition> topicPartitions = kafkaConsumer.listTopics().get(KAFKA_TOPIC).stream().map(partitionInfo -> new TopicPartition(partitionInfo.topic(), partitionInfo.partition())).collect(Collectors.toList());
+        Map<TopicPartition, Long> topicPartitionOffsets = kafkaConsumer.endOffsets(topicPartitions);
+        topicPartitionOffsets.forEach((topicPartition, offset) -> assertThat(channelPosition.shard("" + topicPartition.partition()).position(), is("" + (offset - 1))));
     }
 
     @Configuration